Rename nesting to nested
In bug 1306708, billm said: "I sort of settled on "nested level" as a replacement for priority. I think "nested" is a little better than "nesting" because it makes it clear that it's about what the message can nest in rather than what can nest inside it."

Rename the fields to match the IPDL compiler.