Clarify the preproc hack's nature, test preproc in comments.
This addresses post-landing review action items from
https://github.com/mozsearch/ipdl_parser/pull/12

Note that I've lined up the assertion strings so it's visually easier
to validate what's going on in the code but I'm not sure a code
formatter would be happy with it.

Also note that the `error_tests` test case is unhappy but that is
unrelated to these changes.